Split StaticCaller to separate file for PSR-4 compliance
